LPS-165445 Merge DefaultIndexerRequestBufferOverflowHandler and IndexerRequestBufferOverflowHandler
